backend/verify_temp.py
import requests
import json
import sys

BASE_URL = "http://127.0.0.1:8000/api/v1"
EMAIL = "test@example.com"
PASSWORD = "password123"

def print_step(msg):
    print(f"\n[STEP] {msg}")

def run_verification():
    # 1. Register/Login (MVP: Create user via shell if needed, or assume pre-created)
    # We will automate creating a user via Django shell for this script to work
    
    # 2. Login to get Token
    print_step("Obtaining JWT Token...")
    try:
        response = requests.post(f"{BASE_URL}/auth/login/", data={"email": EMAIL, "password": PASSWORD})
        # If auth endpoint not ready, we skip authentication for testing by temporarily disabling it in views?
        # A better way for this script: Use manage.py shell to create data directly first to verify DB.
        # But we want to test APIs.
        pass
    except Exception as e:
        print(f"Error: {e}")

if __name__ == "__main__":
    print("This script is a template. We'll use a Django Management Command instead for direct DB access verification.")
