backend/requirements/base.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
Pillow>=10.0
python-decouple>=3.8
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
drf-spectacular>=0.27
django-storages>=1.14
cloudinary>=1.36
sentry-sdk>=1.40
channels>=4.0
channels-redis>=4.1
